+++ ufotest/tests/basic.py
@@ -1,0 +1,12 @@
+from ufotest.testing import AbstractTest, MessageTestResult, AssertionTestResult, TestRunner
+
+
+class RepeatedResetTest(AbstractTest):
+
+    name = 'repeated_setup'
+    description = (
+        'this is a description'
+    )
+
+    def run(self):
+        pass