+++ ufotest/tests/mock.py
@@ -1,6 +1,7 @@
-import time
+import random
 
 import numpy as np
+import matplotlib.pyplot as plt
 from PIL import Image
 
 from ufotest.util import random_string
@@ -9,7 +10,8 @@ from ufotest.testing import (AbstractTest,
                              MessageTestResult,
                              ImageTestResult,
                              CombinedTestResult,
-                             DictTestResult)
+                             DictTestResult,
+                             FigureTestResult)
 
 
 class MockTest(AbstractTest):
@@ -28,18 +30,18 @@ class MockTest(AbstractTest):
     def run(self):
         exit_code = 0
 
-        # -- MESSAGE RESULT
+        # ~ MESSAGE RESULT
         message = 'Mock test case message'
         message_test_result = MessageTestResult(exit_code, message)
 
-        # -- IMAGE RESULT
+        # ~ IMAGE RESULT
         array = np.random.randint(0, 65535, size=(480, 720)).astype(np.uint16)
         image = Image.fromarray(array)
-        image_path = '{}/{}.png'.format(self.context.folder_path, random_string(10))
+        image_path = self.context.get_path(f'{random_string(10, additional_letters="")}.png')
         image.save(image_path)
         image_test_result = ImageTestResult(exit_code, image_path, 'A random image.', self.context.folder_url)
 
-        # -- DICT RESULT
+        # ~ DICT RESULT
         table = {
             'key1':         'value1',
             'key2':         'value2',
@@ -47,8 +49,28 @@ class MockTest(AbstractTest):
         }
         dict_test_result = DictTestResult(exit_code, table)
 
+        # ~ FIGURE RESULT
+        fig, (ax1, ax2) = plt.subplots(nrows=1, ncols=2, figsize=(8, 8))
+
+        r = np.random.randn(100)
+        r1 = r + 2
+        ax1.hist([r, r1])
+        ax2.hist([r1, r])
+
+        ax1.set_title('Random value distribution')
+        ax1.set_xlabel('x axis')
+        ax1.set_ylabel('y axis')
+
+        figure_test_result = FigureTestResult(
+            exit_code=exit_code,
+            test_context=self.context,
+            figure=fig,
+            description='A matplotlib figure as the test result'
+        )
+
         return CombinedTestResult(
             message_test_result,
             image_test_result,
-            dict_test_result
+            dict_test_result,
+            figure_test_result,
         )